+++ src/ambition/version_details.cc
@@ -30,6 +30,7 @@ std::vector<VersionUpdates> versionDetails = {
           _("Fixed the building expense amounts on the Villages report being lower than the real expense."),
           _("Fixed spy icon getting duplicated when workers are re-arranged inside buildings."),
           _("Fixed raw material (and other) icons showing with a white background."),
+          _("Fixed potential incorrect placement and crash with user interface bounds."),
         },
       },
       {